--- go.mod
@@ -1,3 +1,0 @@
-module myapp
-
-go 1.24.3
