+++ src/go.mod
@@ -1,0 +1,3 @@
+module mainapp
+
+go 1.24.3
