+++ src/pkg/models/answer.go
@@ -1,0 +1,7 @@
+package models
+
+
+type Answer struct {
+	Id          uint32
+	Description string
+}
