+++ src/pkg/models/poll.go
@@ -1,0 +1,7 @@
+package models
+
+type Poll struct {
+	Id        uint32
+	Title     string
+	Questions []QuestionInterface
+}
