+++ src/pkg/models/submission.go
@@ -1,0 +1,5 @@
+package models
+
+type Submission struct {
+	Id uint32
+}
